+++ src/sections/userpreset-browser.ts
@@ -228,21 +228,12 @@ export class UserPresetBrowser extends FavBrowserBase {
       this.alertInfo = "Playing recommended tracks ...";
       this.requestUpdate();
 
-      // if shuffle enabled then disable it, as we want to play the selected track first.
-      if (this.player.attributes.shuffle == true) {
-        await this.store.mediaControlService.shuffle_set(this.player, false);
-      }
+      // play the selected track, as well as the remaining tracks.
+      // also disable shuffle, as we want to play the selected track first.
+      this.spotifyPlusService.PlayerMediaPlayTracks(this.player, uris.join(","), null, null, null, false);
 
-      // give the shuffle disable time to process.
-      setTimeout(() => {
-
-        // play recommended tracks.
-        this.spotifyPlusService.PlayerMediaPlayTracks(this.player, uris.join(","));
-
-        // show player section.
-        this.store.card.SetSection(Section.PLAYER);
-
-      }, 250);
+      // show player section.
+      this.store.card.SetSection(Section.PLAYER);
 
     }
     catch (error) {